Refactor creation of codec maps into lib/codec.js.
