+++ lib/decode.js
@@ -43,7 +43,9 @@ export async function fromCborld({
   } else if(cborMap.tag === 0x0501) {
     const contextUrls =
       getCborldContextUrls({cborMap: cborMap.value, appContextMap});
-    //codecMap = getCodecMap({contextUrls, appTermMap});
+    const codecMap = getCodecMap({contextUrls, appTermMap});
+
+    console.log("CODEC MAP", codecMap);
   } else {
     throw new CborldError('ERR_NOT_CBORLD',
       'Input is not valid CBOR-LD; missing CBOR-LD header (0x0500 or 0x501).');
